+++ lib/Perl/Tidy/DevNull.pm
@@ -7,7 +7,7 @@
 package Perl::Tidy::DevNull;
 use strict;
 use warnings;
-our $VERSION = '20221112';
+our $VERSION = '20221112.01';
 sub new   { my $self = shift; return bless {}, $self }
 sub print { return }
 sub close { return }