+++ lib/Perl/Tidy/LineBuffer.pm
@@ -12,7 +12,7 @@
 package Perl::Tidy::LineBuffer;
 use strict;
 use warnings;
-our $VERSION = '20221112';
+our $VERSION = '20221112.01';
 
 sub AUTOLOAD {