+++ lib/Perl/Tidy/LineSource.pm
@@ -9,7 +9,7 @@ package Perl::Tidy::LineSource;
 use strict;
 use warnings;
 use English qw( -no_match_vars );
-our $VERSION = '20221112';
+our $VERSION = '20221112.01';
 
 use constant DEVEL_MODE => 0;